Rename sub directory for installation stores
